+++ CHANGELOG.md
@@ -23,7 +23,7 @@
 - Added the ability to run targets in `moon run` and `moon ci` using a glob-like syntax.
   - For example: `:build-*`, `app-*:build`, `#tag-{foo,bar}:build`, etc.
 - Added a `--no-bail` flag to `moon run`, that will continue running tasks even when a task fails.
-- Added `glob://` locator support to `generator.templates`, allowing you to glob for your codegen
+- Added glob support (and `glob://`) to `generator.templates`, allowing you to glob for your codegen
   templates locations.
 - Updated task option `runInCI` to support the values "always" (always run) and "affected" (only run
   if affected, same as `true`).